requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.3
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
idna==3.2
line-bot-sdk==1.19.0
platformdirs==2.0.2
psycopg2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.6.0
whitenoise==5.3.0
